requirements.txt
altair==4.2.0
anyio==3.6.1
aplus==0.11.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astropy==5.1
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
blake3==0.3.1
bleach==4.1.0
blinker==1.4
bqplot==0.12.33
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.1.0
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
dask==2022.7.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.78.0
filelock==3.7.1
fonttools==4.29.1
frozendict==2.3.2
fsspec==2022.5.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
h5py==3.7.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
ipydatawidgets==4.3.1.post1
ipykernel==6.9.1
ipyleaflet==0.17.0
ipympl==0.9.1
ipython==7.32.0
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipyvue==1.7.0
ipyvuetify==1.8.2
ipywebrtc==0.6.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
llvmlite==0.38.1
locket==1.0.0
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numba==0.55.2
numpy==1.21.5
openai==0.27.0.
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pandas-stubs==1.5.1.221024
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
progressbar2==4.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pyarrow==7.0.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pydeck==0.7.1
pyerfa==2.0.0.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-utils==3.3.3
pythreejs==2.3.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.2.0
requests==2.27.1
rich==12.4.4
scikit-learn==1.0.2
scipy==1.7.3
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.19.1
streamlit==1.20.0
tabulate==0.8.10
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
traittypes==0.2.1
types-pytz==2022.6.0.1
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.18.2
uvloop==0.16.0
vaex==4.9.2
vaex-astro==0.9.1
vaex-core==4.9.2
vaex-hdf5==0.12.2
vaex-jupyter==0.8.0
vaex-ml==0.17.0
vaex-server==0.8.1
vaex-viz==0.5.2
validators==0.18.2
watchdog==2.1.6
watchfiles==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
widgetsnbextension==3.5.2
xarray==2022.3.0
xyzservices==2022.6.0
zipp==3.7.0
